Updated neural_model.py
Removed old neural_model.py
Added updated neural_model.py
